% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhSDM_notexported.R
\name{gRasterize}
\alias{gRasterize}
\title{Rasterize features}
\usage{
gRasterize(spf, rast, value = 1, background = NA)
}
\arguments{
\item{spf}{input sf or sp object}

\item{rast}{terra raster dataset with desired output projection, extent, cell size}

\item{value}{integer value to apply to areas covered by spf}

\item{background}{value to apply to areas not covered by spf}
}
\value{
rast object
}
\description{
This is a wrapper around terra::rasterize, used internally in nh_burn and nh_stack.
}
\keyword{internal}
