% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_best.R
\name{nh_best}
\alias{nh_best}
\title{Extract areas with best (highest value) as polygons from an SDM prediction raster, with optional feature mask}
\usage{
nh_best(
  rast,
  spf = NULL,
  top.percent = NULL,
  min.size = NULL,
  min.dist = NULL,
  num.patches = NULL,
  rank.by = "area"
)
}
\arguments{
\item{rast}{input raster model output with continuous values}

\item{spf}{input spatial features (sp or sf spatial object); if given, it can be used to modify areas selected from \code{rast}}

\item{top.percent}{numeric; percent (e.g.; 0.01 = 0.01\%) of highest cell values in raster to extract}

\item{min.size}{numeric; optional minimal area of an extracted polygon}

\item{min.dist}{numeric; optional minimal distance from spf for patches. Default is 0 (can be adjacent to 'spf')}

\item{num.patches}{numeric; optional number of patches to return, using rank.by criteria}

\item{rank.by}{character; used for ranking patches when num.patches is not null; either 'area' (default) or 'value'}
}
\value{
sp or sf object (polygons)
}
\description{
If \code{spf} is given, areas intersecting these features (plus a buffer, if \code{min.dist} is specified) 
will not be included in the returned polygons.
}
\details{
\code{top.percent} and \code{min.size} will both be derived from \code{spf} if they are null
and \code{spf} is given. In this case, \code{top.percent} will be set to be equal to \code{spf}'s cell coverage 
(prevalence relative to non-NA areas in raster). \code{min.size} will be set to the area 
of the smallest feature in \code{spf}. If \code{spf} are line or point features, this will be zero. 

\code{rank.by} only subsets the outputs if \code{num.patches} is set. \code{rank.by}
must be set to \code{"value"} to return the mean prediction value of the patch.

For better performance, crop the raster prior to running, and/or use a very low 
\code{top.percent} value, to return a smaller proportion of the cells as polygons.
}
\examples{
\dontrun{
spf <- sf::st_read("_data/occurrence/ambymabe.shp")
rast <- terra::rast("_data/species/ambymabe/outputs/model_predictions/ambymabe_20171018_130837.tif")
rast <- terra::crop(rast, spf)
best_polys <- nh_best(rast, spf, top.percent = 0.01, min.size = 10000, min.dist <- 10000, 
   num.patches = 100, rank.by = "value")
}
}
\author{
David Bucklin
}
