% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_burn.R
\name{nh_burn}
\alias{nh_burn}
\title{Add areas represented by features ('burn-in') to a binary raster SDM output}
\usage{
nh_burn(spf, rast, orig.thresh = NULL, buffer = 0, ...)
}
\arguments{
\item{spf}{input spatial features (sp or sf spatial object)}

\item{rast}{input raster model output (values between 0 and 1)}

\item{orig.thresh}{numeric between 0 and 1; threshold value to apply to full raster}

\item{buffer}{numeric; spatial buffer around spf to include in burn-in}

\item{...}{additional parameters to \code{raster::writeRaster}}
}
\value{
RasterLayer
}
\description{
Takes spatial features, a continuous raster (values between 0 and 1), and 
returns a classified binary (0/1) raster. Areas greater than \code{orig.thresh} are
set to 1, all others 0. Areas intersecting \code{spf} features
are assigned a value of 1 in the returned classified raster. See details
for default calculation of \code{orig.thresh} and usage of \code{buffer}.
}
\details{
When \code{buffer} is used or \code{orig.thresh} is not provided,
a minimum cell value (min.cell) across all \code{spf} is calculated.
The min.cell value will be used as a threshold
for the full raster, or just areas within \code{buffer} distance of \code{spf},
if \code{buffer} is greater than 0.

If \code{buffer} is greater than 0 \code{orig.thresh} is not NULL, \code{orig.thresh}
is used as the threshold for the full raster. Additionally,
cells within \code{buffer} distance of \code{spf} which are greater than min.cell
are also set to 1. If a given \code{orig.thresh} is lower 
than the calculated min.cell value, \code{buffer}
distance will have no impact on the output.
}
\examples{
\dontrun{
spf <- st_read("ambymabe/polygon_data/ambymabe.shp")
rast <- raster("ambymabe/grids/ambymabe_20171018_130837.tif")
class_burn <- nh_burn(spf, rast, 0.75, 250, filename = "model_classified.tif", datatype = "INT2U")
}
}
\author{
David Bucklin
}
