% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_group.R
\name{nh_group}
\alias{nh_group}
\title{Group spatial features using a defined separation distance}
\usage{
nh_group(spf, sep.dist = 0, union = FALSE)
}
\arguments{
\item{spf}{input spatial features (sp or sf spatial object)}

\item{sep.dist}{separation distance with which to define groups (see description)}

\item{union}{whether to union output groups into multi-features}
}
\description{
Input features are allocated into new groups depending
on whether they lie within the separation distance of
another input feature.
}
\details{
The grouping is done by input ID, so multi-features
are considered one input feature (they will not be
be ungrouped).

The separation distance \code{sep.dist} is numeric and in the
units of \code{spf}'s coordinate system, unless the
coordinate system uses latitude/longitude as the unit (e.g. WGS 84). In
these cases, geodesic distances will be used and \code{sep.dist}
should be specified in meters.

A column 'group' will be added to the output features.
Specifying \code{union = TRUE} will output one (multi)feature per group,
meaning original attributes will be discarded - only `group` and `count`
(the number of original features in the group) will be returned. This feature
requires the package \code{dplyr} to be installed.
}
\examples{
\dontrun{
spf <- rgdal::readOGR("ambymabe/polygon_data", "ambymabe")
spg <- nh_group(spf, 1000)
}
}
\author{
David Bucklin
}
