% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_linefill.R
\name{nh_linefill}
\alias{nh_linefill}
\title{For line-based SDM predictions, identify areas between predicted suitable
lines}
\usage{
nh_linefill(spf, field, max.dist = NA, max.line = NA)
}
\arguments{
\item{spf}{input spatial features (model predictions; sp or sf spatial object)}

\item{field}{name of field in \code{spf} (binary 1/0), where lines will be filled between features with value == 1}

\item{max.dist}{maximum distance between selected lines to fill in}

\item{max.line}{maximum number of lines between selected lines to fill in}
}
\value{
sf object including suitable lines and filled-in lines
}
\description{
Provided a network of lines (e.g., reaches in a hydrological network), and 
an attribute that identifies those lines representing suitable habitat, this
function identifies fill-in lines between suitable lines, given certain limits
(total distance and/or total reaches). It does not alter original line segments.
}
\details{
Line directionality is used to identify how they are connected in the network,
so start/end nodes must align. These nodes will be calculated; alternatively, 
the names \code{startNode} and \code{endNode} can be provided as columns in \code{spf} and 
will be used instead.

Values for \code{max.dist} should be in the units of the CRS, except when
using a lat/lon based CRS, in which case the value should be in meters.
}
\examples{
\dontrun{
spf <- st_read("SDM_results.shp")
fill <- nh_linefill(spf, "thresh", max.dist = 5000, max.line = 5)
}
}
\author{
David Bucklin
}
