% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_map.R
\name{nh_map}
\alias{nh_map}
\title{Make a suitable/unsuitable binary map, given a continuous SDM output or raster template}
\usage{
nh_map(rast, thresh = NULL, feature.occ = NULL, feature.burn = NULL,
  feature.mask = NULL, raster.mask = NULL, patch.drop = NULL)
}
\arguments{
\item{rast}{input raster model output (values between 0 and 1); if no model, a template raster must be provided}

\item{thresh}{numeric between 0 and 1; 'global' threshold value to apply to rast}

\item{feature.occ}{input feature occurrences, to burn-in to map (sp or sf spatial object)}

\item{feature.burn}{input feature(s), to burn-in to map (sp or sf spatial object)}

\item{feature.mask}{input feature mask to apply to map (sp or sf spatial object)}

\item{raster.mask}{input raster mask to apply to map (should match \code{rast} resolution)}

\item{patch.drop}{patch size (in \code{rast} area units) to remove from map}
}
\value{
RasterLayer
}
\description{
This function performs a reclassification of a continuous map to a binary one,
and then performs additional changes to the binary map, depending on inputs. If
no continuous map is to be used, then a template raster should be given to \code{rast}.
}
\details{
Reclassification (when \code{thresh} is given) is always peformed first. 
After this, optional masks are applied where areas within features (\code{feature.mask}) 
and not NoData (\code{raster.mask}) are kept (value = 1). All other cells are given (value = 0).

Patches (contiguous groups of cells = 1) are then be dropped by specifying a 
minimum patch size to \code{patch.drop} in area units of the raster's CRS 
(see nhSDM::nh_patchdrop).

The final step is to 'burn in' (rasterize with value = 1) features 
from \code{feature.occ} and \code{feature.burn}. Since this is the last step, it is not
affected by masks or patch dropping.
}
\examples{
\dontrun{
spf <- st_read("ambymabe/polygon_data/ambymabe.shp")
rast <- raster("ambymabe/grids/ambymabe_20171018_130837.tif")
map <- nh_map(rast, 0.75, feature.occ = spf)
}
}
\author{
David Bucklin
}
