% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_proj.R
\name{nh_newproj}
\alias{nh_newproj}
\title{Initiate new SDM project}
\usage{
nh_newproj(proj.name, folder = "nhSDM")
}
\arguments{
\item{proj.name}{The name of the project. Appended to a new project database file.}

\item{folder}{The SDM folder in which to initiate the project. Defaults to `getwd()`.}
}
\value{
TRUE on succesful project creation
}
\description{
Creates directory structure (if it doesn't exist) for SDMs and 
new SDM project SQLite database. Folder structure is initiated in current
working directory, unless \code{folder} is specified.
}
\details{
More details
}
\examples{
\dontrun{
setwd("D:/")
nh_newproj("invert")
}
}
\author{
David Bucklin
}
