% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_patchdrop.R
\name{nh_patchdrop}
\alias{nh_patchdrop}
\title{Remove contiguous patches smaller than a given patch size from binary output}
\usage{
nh_patchdrop(spf = NULL, rast, min.patch = NULL, directions = 8, ...)
}
\arguments{
\item{spf}{input spatial features (sp or sf spatial object)}

\item{rast}{input binary raster output (values either NA/0 or 1)}

\item{min.patch}{area of minimum patch size, in area units used in \code{rast}}

\item{directions}{Integer. Which cells are considered adjacent? Should be 8 (default; Queen's case) or 4 (Rook's case). From \code{raster::clump}}

\item{...}{Other arguments as to \code{raster::writeRaster}}
}
\value{
RasterLayer
}
\description{
Takes a binary/thresholded raster (values either NA, 0, or 1), and 
returns a binary raster. Clumps of contiguous cells with the value (1) 
that are smaller than the \code{min.patch} size are given a value of 0 
(if 0 values are present in the input 'rast'); otherwise, they
are given an NA value.
}
\details{
If 'spf' is given, the smallest feature's area will be used to derive
a \code{min.patch} value, and any given 'min.patch' is ignored. If 'spf' is 
not given, a 'min.patch' value must be given, in area units of the input raster.
}
\examples{
\dontrun{
spf <- st_read("ambymabe/polygon_data/ambymabe.shp")
rast <- raster("screen_lpsh/thumb/ambymabe.tif")

# use minimum patch size from presence features
rast_contig_minpres <- nh_patchdrop(spf, rast)

# use a minimum patch size of 10000 (in units from 'rast')
rast_contig_10km <- nh_patchdrop(rast = rast, min.patch = 10000,
 filename = "rast_contig_10km.tif", datatype = "INT2U")
}
}
\author{
David Bucklin
}
