% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_proj.R
\name{nh_proj}
\alias{nh_proj}
\title{Initiate new SDM project}
\usage{
nh_proj(proj.name, folder = ".")
}
\arguments{
\item{proj.name}{The name of the project. New folder created and name appended to a new project database file.}

\item{folder}{The folder in which to initiate the project}
}
\value{
nothing
}
\description{
Creates directory structure for a new SDM project and 
new SDM project SQLite database. New project folder is 
created in working directory, unless \code{folder} is specified.
}
\examples{
\dontrun{
setwd("D:/testing_SDM")
nh_proj("new_proj")
}
}
\author{
David Bucklin
}
