% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_stack.R
\name{nh_stack_resample}
\alias{nh_stack_resample}
\title{Aggregate values from a nh_stack to a lower (coarser) resolution raster, or within polygons}
\usage{
nh_stack_resample(rast, lookup, fact = 10, spf = NULL)
}
\arguments{
\item{rast}{raster output from nh_stack}

\item{lookup}{lookup table from nh_stack}

\item{fact}{aggregation factor, in number of cells (see \code{?terra::aggregate})}

\item{spf}{Optional vector spatial features to use for aggregation (sp or sf-class polygons). If supplied, \code{fact} will be ignored}
}
\value{
SpatRaster or spf (with attributes added)
}
\description{
When, \code{spf=NULL}, takes an output raster from nh_stack, and returns a lower-resolution
version, with recalculated species assemblages for the larger cells.
New values are "aggregated" by \code{fact}, the number of cells
to aggregate in the x/y dimensions (see \code{?terra::aggregate}).
}
\details{
Alternatively, you can aggregate species assemblages within polygons (sp or sf-class) 
provided to \code{spf}. Polygons intersecting areas with non-NA
cells in \code{rast} are returned, with columns identifying species codes 
and counts. Polygons will be returned in their original projection, but 
processing internally is done in the raster's projection.
}
\examples{
\dontrun{
# stack <- nh_stack(list, rast, return.table = TRUE)

# resample from 30m to 990m resolution
stack1km <- nh_stack_resample(stack[[1]], stack[[2]], fact = 33)
# view raster attributes
cats(stack1km)

# view species count raster (index=3 is the 'ALLCODES_CT' column)
ct <- as.numeric(stack1km, index=3)
plot(ct)
}
}
\author{
David Bucklin
}
