% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhSDM_notexported.R
\name{nodes}
\alias{nodes}
\title{Calculate nodes in a line network, using line direction}
\usage{
nodes(spf)
}
\arguments{
\item{spf}{input sf object}
}
\value{
sf object
}
\description{
Returns same object with added columns startNode and endNode
}
\keyword{internal}
