% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_group.R
\name{poly_group}
\alias{poly_group}
\title{Group spatial polygons using a defined separation distance}
\usage{
poly_group(spP, sep.dist, union = FALSE)
}
\arguments{
\item{spP}{input SpatialPolygonsDataFrame}

\item{sep.dist}{separation distance (in projection units) with which to define groups}

\item{union}{whether to union output groups in the resulting SPDF}
}
\description{
Input polygons are allocated into new groups depending
on whether they lie within the seperation distance of
another input polygon.
}
\details{
The grouping is done by input ID, so multi-polygons
are considered as one input polygon (they will not be
be ungrouped.) Grouping is greedy - for a given input
polygon, it will belong to the group containing all of
its neighbor polygons, plus all of their neighboring
polygons, etc.

A column 'group' will be added to the output SpatialPolygonsDataFrame.
Specifying \code{union = TRUE} will output one (multi)polygon per group,
meaning original attributes will be discarded.
}
\author{
David Bucklin
}
