% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_samp.R
\name{poly_samp}
\alias{poly_samp}
\title{Create points in polygons in unique raster cells,
For each spatial polygon, a given number (\code{num_samps})
of points are created in cells that the polygon intersects,
with only one point allowed per cell (in the polygon and the
entire dataset).}
\usage{
poly_samp(spP, rast, num.samps = NULL)
}
\arguments{
\item{spP}{SpatialPolygons[DataFrame]}

\item{rast}{raster dataset with extent greater or equal to spP}

\item{num.samps}{number of samples to create in each polygon}
}
\description{
\code{num.samps} can be a single integer, or vector of integers 
equal to length of spP. If left NULL, the number of samples per
polygon will be equal to {[area of polygon]/[area of raster cell] * 5}.
If the number of samples to create exceeds the number of cells in a given
polygon, the number of samples is set to the number of cells.
}
\examples{
\dontrun{
a<-raster("D:/SDM/Tobacco/env_vars/Tobacco/AnnMnTemp.tif")
spP <- readOGR("D:/SDM/Tobacco/inputs/species/ambymabe/polygon_data", "ambymabe_expl")
s <- c(4,3,5)[spP$RA]
sp <- poly_samp(spP, a, s)
}
}
\author{
David Bucklin
}
