% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_group.R
\name{shiny_group}
\alias{shiny_group}
\title{Interactively view (in a leaflet map) and select grouping distance}
\usage{
shiny_group(spP)
}
\arguments{
\item{spP}{A SpatialPolygons* object}
}
\description{
Interactively view (in a leaflet map) and select grouping distance
}
\examples{
\dontrun{
spP <- readOGR("D:/SDM/Tobacco/inputs/species/ambymabe/polygon_data", "ambymabe_expl")
shiny_group(spP)
}
}
