% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhSDM_notexported.R
\name{tospf}
\alias{tospf}
\title{Return a list with sp-flag and sf object, given either sf or sp object}
\usage{
tospf(spf, rastproj)
}
\arguments{
\item{spf}{input sf or sp object}

\item{rastproj}{raster dataset with desired output projection}
}
\description{
Return a list with sp-flag and sf object, given either sf or sp object
}
\keyword{internal}
