% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_burn.R
\name{nh_burn}
\alias{nh_burn}
\title{Add areas represented by features ('burn-in') to a binary raster SDM output, 
plus areas within a buffer distance above a threshold}
\usage{
nh_burn(spf, rast, orig.thresh = NULL, buffer = NA, return.thresh = FALSE)
}
\arguments{
\item{spf}{input spatial features (sp or sf spatial object)}

\item{rast}{input raster model output (values between 0 and 1)}

\item{orig.thresh}{numeric between 0 and 1; threshold value to apply to raster}

\item{buffer}{numeric; spatial buffer distance around spf where threshold will be applied.}

\item{return.thresh}{logical; whether to return thresholds along with raster in a list}
}
\value{
SpatRaster
}
\description{
Takes spatial features, a continuous raster (values between 0 and 1), and 
returns a classified binary (0/1) raster. Areas intersecting \code{spf} features
are assigned a value of 1 in the returned classified raster. Additionally, areas
within \code{buffer} distance of \code{spf} and above the threshold are set to 1. 
See details for default calculation of \code{orig.thresh} and usage of \code{buffer}.
}
\details{
When \code{buffer} is used or \code{orig.thresh} is not provided,
a minimum cell value (min.cell) of values within \code{spf} is calculated and 
used as the threshold.

The default is to return the raster only. If \code{return.thresh = TRUE}, the function will
return a list with 3 named objects: \code{rast}, the output raster; \code{orig.thresh}, 
the global threshold (if used); \code{min.cell}, the minimum cell value threshold (if used).
}
\examples{
\dontrun{
spf <- sf::st_read("_data/occurrence/ambymabe.shp")
rast <- terra::rast("_data/species/ambymabe/outputs/model_predictions/ambymabe_20171018_130837.tif")
full_burn <- nh_burn(spf, rast, buffer = NA)
buff_burn <- nh_burn(spf, rast, buffer = 10000)
}
}
\author{
David Bucklin
}
