% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_burn.R
\name{nh_burn}
\alias{nh_burn}
\title{Add areas represented by features ('burn-in') to a binary raster SDM output}
\usage{
nh_burn(spf, rast, orig.thresh = NULL, buffer = 0, ...)
}
\arguments{
\item{spf}{input spatial features (sp or sf spatial object)}

\item{rast}{input raster model output (values between 0 and 1)}

\item{orig.thresh}{numeric between 0 and 1; threshold value to apply to raster}

\item{buffer}{numeric; spatial buffer around spf to include in burn-in}

\item{...}{additional paramaters to \code{raster::writeRaster}}
}
\description{
Takes spatial features, a continuous raster (values between 0 and 1), and 
returns a classified binary (0/1) raster. Areas greater than \code{orig.thresh} are
set to 1, all others 0. Areas intersecting \code{spf} features
are assigned a value of 1 in the returned classified raster.
}
\details{
If \code{buffer} is not 0, areas within that buffer around \code{spf} that are greater 
than the minimum cell value (min.cell) within \code{spf} are set to 1 
in the classified raster.

If \code{orig.thresh} is NULL, the min.cell value will be used as a global threshold.
When a given \code{orig.thresh} is less than the min.cell value, the result
will be the same as a thresholded raster using \code{orig.thresh}.
}
\examples{
\dontrun{
spf <- st_read("ambymabe/polygon_data/ambymabe.shp")
rast <- raster("ambymabe/grids/ambymabe_20171018_130837.tif")
class_burn <- nh_burn(spf, rast, 0.75, 250, filename = "model_classified.tif", datatype = "INT2U")
}
}
\author{
David Bucklin
}
