% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_rasterize.R
\name{nh_rasterize}
\alias{nh_rasterize}
\title{Convert vector format SDM prections to raster format}
\usage{
nh_rasterize(spf, rast, pred.vals, buffer = 0, priority = NULL,
  rast.out = NULL, ...)
}
\arguments{
\item{spf}{input spatial features (model predictions; sp or sf spatial object)}

\item{rast}{input raster template}

\item{pred.vals}{prediction values for spf features}

\item{buffer}{numeric; spatial buffer around spf to include in burn-in}

\item{priority}{numeric; vector of priority values for feature rasterization (higher values have priority)}

\item{rast.out}{Output raster file name (with file extension)}

\item{...}{Additional arguments to writeRaster (e.g. overwrite)}
}
\value{
RasterLayer
}
\description{
Requires spatial features, a template raster, and prediction
values vector. The buffer is optional; when provided, the
buffer distance can be provided as a single numeric value or a 
vector matching length of \code{spf}, for a variable buffer by feature.
}
\details{
Buffer units should be given in the units of \code{rast}. Make 
sure to keep in mind the resolution of the raster when choosing 
a buffer. If rasterizing lines and a one-cell width is desired,
do not use a buffer.

A vector can be given to `priority` for sorting prior to rasterization, 
where higher values have priority. When \code{priority = NULL} (default),
priority will be defined as the \code{pred.vals}.

If \code{rast.out} is not specified, the raster will remain in temp folder.
}
\examples{
\dontrun{
spf <- st_read("acipoxyr/shapefiles/acipoxyr_20180105_133929_results.shp")
rast <- raster("template.tif")

# rasterize
bla <- nh_rasterize(spf, rast, pred.vals = spf$prbblty, buffer = spf$strord*15)
}
}
\author{
David Bucklin
}
