% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_sample.R
\name{nh_sample}
\alias{nh_sample}
\title{Create points in polygons in reference raster cells}
\usage{
nh_sample(spf, rast, num.samps = NULL, replace = FALSE, force.min = FALSE)
}
\arguments{
\item{spf}{input spatial features (sp or sf spatial object)}

\item{rast}{raster dataset with extent overlapping spf}

\item{num.samps}{number of samples to create in each polygon (see details)}

\item{replace}{whether to sample with or without replacement}

\item{force.min}{whether to always create \code{num.samps} points, even if they are duplicates}
}
\description{
For each spatial polygon, a given number (\code{num_samps})
of points are created in cells that the polygon intersects.
}
\details{
\code{num.samps} can be a a proportion (a decimal value < 1), single integer,
or vector of integers equal to length of spf indicated the number of samples
to take from each polygon. If left NULL, \code{num.samps} will
be set to the number of cells [n] intersecting the polygon. If a proportion is given
(e.g., 0.5), than [n * num.samps] will be returned. If a single integer is given,
\code{num.samps} points will be sampled in each polygon.

When \code{replace = FALSE} and {force.min = FALSE} (defaults),
each cell can only contain one point (across
each polygon, and the entire returned set of points). In this case, when the number
of samples points to create exceeds the number of unique cells intersected by a given
polygon, the number of samples for that polygon is reduced to the number of cells.
If \code{replace = TRUE}, sampling is done with replacement and duplicates
may be taken. The special case \code{replace = FALSE} and \code{force.min = TRUE}
will always return \code{num.samps} per polygon. It only produces duplicates
if \code{num.samps} exceeds the number of cells intersecting the polygon,
in which case it will replicate the samples until \code{num.samps} is reached.

If CRS do not match, the SpatialPolygons will be transformed to the CRS
of the raster.
}
\examples{
\dontrun{
r<-raster::raster("D:/SDM/Tobacco/env_vars/Tobacco/AnnMnTemp.tif")
spf <- rgdal::readOGR("D:/SDM/Tobacco/inputs/species/ambymabe/polygon_data", "ambymabe")
spg <- nh_sample(spf, 1000, TRUE)
rast <- r
num.samps <- 10
replace <- FALSE
force.min <- FALSE
sp.sing <- nh_sample(spf, r, num.samps)
sp.mult <- nh_sample(spg, r, num.samps)
}
}
\author{
David Bucklin
}
