% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_stack.R
\name{nh_stack}
\alias{nh_stack}
\title{Stack multiple binary SDM rasters into one raster layer}
\usage{
nh_stack(rastfiles, rast, codes = NULL, return.table = TRUE)
}
\arguments{
\item{rastfiles}{raster file names; character vector}

\item{rast}{raster template dataset}

\item{codes}{species codes; character vector. If given, must match length of \code{rastfiles}}

\item{return.table}{Whether to return a table with nh_stack unique values, species codes, and filenames}
}
\value{
RasterLayer
}
\description{
Takes a \code{rastfiles} of binary raster's filenames (values 0/1), 
a template raster covering the extent desired for the stack, 
and optionally the \code{codes} (names) to use
for the rasters. Returns a single raster layer (factor type with an
attribute table).
}
\details{
The raster attribute table (accessed using \code{levels()}) 
has four columns:  `ID`: the unique integer value in the raster; `VALUE`:
the internal nh_stack unique value for that `ID`, `ALLCODES`: the
identity of species codes, pasted in a character vector seperated by `;`,
`ALLCODES_CT`: the number of unique codes for that value.

If \code{codes} is not given, the raster layer name will be used as the 
layer's code. If these are not unique, the internal nh_stack unique value will
be pasted to the end of the original code, and a message will be printed.

All rasters must have the same projection and resolution, though
they can have different extents - the processing extent is defined by
\code{rast}. To summarize all rasters across their
entire extents, \code{template} should essentially be the union (mosaic) 
of all raster extents.

When \code{return.table} is TRUE, a list with two objects (1), the 
stack raster, and (2) a summary table of included rasters is returned (with
internal nh_stack unique values, species codes, and file names). This table
is required for resampling the stack (i.e. with \code{nh_stack_resample}).
}
\examples{
\dontrun{
rast<-raster::raster("project_mask.tif")
list <- sort(list.files(paste0(rastout, t, "/thumb"), full.names = T,
 pattern = "^[[:lower:]]{5,12}\\\\.tif$"))

stack <- nh_stack(list, rast)
# view raster attribute table
levels(stack)[[1]]
}
}
\author{
David Bucklin
}
