% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_stack.R
\name{nh_stack_resample}
\alias{nh_stack_resample}
\title{Resample a raster from nh_stack to a lower (coarser) resolution}
\usage{
nh_stack_resample(rast, lookup, fact = 10)
}
\arguments{
\item{rast}{raster output from nh_stack}

\item{lookup}{lookup table from nh_stack}

\item{fact}{aggregation factor, in number of cells (see \code{?raster::aggregate})}
}
\value{
RasterLayer
}
\description{
Takes an output raster from nh_stack, and returns a lower-resolution
version, with recalculated species assemblages for the larger cells.
New values are "aggregated" by \code{fact}, the number of cells
to aggregate in the x/y dimensions (see \code{raster::aggregate}).
}
\examples{
\dontrun{
stack <- nh_stack(list, rast, return.table = TRUE)

# resample from 30m to 990m (i.e. ~1km) resolution
stack1km <- nh_stack_resample(stack[[1]], stack[[2]], fact = 33)

# view species count raster
ct <- deratify(r2, att = "ALLCODES_CT")
plot(ct)
}
}
\author{
David Bucklin
}
